[Organize] initialize transformation class is added
